+++ R/sm_slope.R
@@ -4,12 +4,9 @@
 #' This function generates a slope chart.
 #' This is very useful for comparing the effect between two time points.
 #'
-#' For this function to work properly,
-#' ggplot()'s mapping has to be quite specific.
-#' 1. Each observation has to be grouped.
+#' ggplot()'s mapping has to be quite specific: each observation has to be grouped.
 #'
-#' With smplot2, average plot can be displayed along with error bars, both of which
-#' are optional.
+#' Error bar types can be specified (ci, sd, and se).
 #'
 #' @param ...
 #' List of parameters for individual points and lines across different elements
@@ -66,8 +63,12 @@
 #' If the legend needs to be displayed, the input should be TRUE.
 #' If the legend is not needed, the input should be FALSE.
 #'
+#' @param forget
+#' Forget the defaults when list() is called for a specific parameter (ex. point.params).
+#' Set to TRUE when when users want to map aesthetics to different groups more flexibly..
+#' Set to FALSE by default.
 #'
-#' @return Returns a double vector output representing the slope from linear regression.
+#' @return Returns a slope chart which is a ggplot2 object.
 #' @import ggplot2 cowplot Hmisc
 #' @importFrom stats sd
 #' @importFrom utils modifyList
@@ -97,7 +98,7 @@
 #'                              alpha = 0.3),
 #'           avgPoint.params = list(color='transparent', shape = 21,
 #'                                  size = 4, fill = sm_color('blue')),
-#'           avgLine.params = list(color = sm_color('blue'), size = 1),
+#'           avgLine.params = list(color = sm_color('blue'), linewidth = 1),
 #'           show_mean = TRUE)
 #'
 #'
@@ -105,13 +106,13 @@ sm_slope <- function(...,
                      labels,
                      group,
                      line.params = list(color = 'gray53',
-                                        size = 0.4,
+                                        linewidth = 0.4,
                                         alpha = 0.4),
                      point.params = list(size = 2.5, shape = 21,
                                          color = 'white'),
-                     avgLine.params = list(),
-                     avgPoint.params = list(),
-                     err.params = list(),
+                     avgLine.params = list(linewidth = 1),
+                     avgPoint.params = list(size = 4),
+                     err.params = list(linewidth = 1),
                      xTick.params = list(position = 'top',
                                          expand = c(0.17,.1),
                                          drop=FALSE),
@@ -119,7 +120,8 @@ sm_slope <- function(...,
                      many_groups = FALSE,
                      show_err = FALSE,
                      show_mean = FALSE,
-                     legends = FALSE) {
+                     legends = FALSE,
+                     forget = FALSE) {
 
   if (missing(group)) {
     stop('group (of the shadow) must be specified because each observation has to be paired.')
@@ -130,12 +132,43 @@ sm_slope <- function(...,
   }
 
   params <- list(...)
-  line.params <- modifyList(params, line.params)
-  point.params <- modifyList(params, point.params)
-  avgLine.params <- modifyList(params, avgLine.params)
-  avgPoint.params <- modifyList(params, avgPoint.params)
-  err.params <- modifyList(params, err.params)
-  #xTick.params <- modifyList(params, xTick.params)
+
+  if (forget == FALSE) {
+    line.params0 = list(color = 'gray53', linewidth = 0.4,
+                        alpha = 0.4)
+    line.params0 <- modifyList(line.params0, params)
+
+    point.params0 = list(size = 2.5, shape = 21,
+                         color = 'white')
+    point.params0 <- modifyList(point.params0, params)
+
+    avgLine.params0 = list(linewidth = 1)
+    avgLine.params0 <- modifyList(avgLine.params0, params)
+
+    avgPoint.params0 = list(size = 4)
+    avgPoint.params0 <- modifyList(avgPoint.params0, params)
+
+    err.params0 = list(linewidth = 1)
+    err.params0 <- modifyList(err.params0, params)
+
+    xTick.params0 = list(position = 'top', expand = c(0.17,.1), drop=FALSE)
+
+
+    line.params <- modifyList(line.params0, line.params)
+    point.params <- modifyList(point.params0, point.params)
+    avgLine.params <- modifyList(avgLine.params0, avgLine.params)
+    avgPoint.params <- modifyList(avgPoint.params0, avgPoint.params)
+    err.params <- modifyList(err.params0, err.params)
+    xTick.params <- modifyList(xTick.params0, xTick.params)
+  } else if (forget == TRUE) {
+    line.params <- modifyList(params, line.params)
+    point.params <- modifyList(params, point.params)
+    avgLine.params <- modifyList(params, avgLine.params)
+    avgPoint.params <- modifyList(params, avgPoint.params)
+    err.params <- modifyList(params, err.params)
+    #xTick.params <- modifyList(params, xTick.params)
+  }
+
 
   if (errorbar_type == 'se') {
     errPlot <- do.call('stat_summary',